# solutions/rsw-reference/RSW.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RSW Official",
    "version": "1.0.0",
    "contact": {}
  },
  "paths": {
...
    "/projects/{projectId}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get project",
        "description": "Get project",
        "operationId": "getProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Soft Delete Project",
        "description": "Soft Delete Project",
        "operationId": "softDeleteProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "patch": {
        "tags": ["projects"],
        "summary": "Update Project",
        "description": "Update Project",
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "New description"
                  },
                  "name": {
                    "type": "string",
                    "example": "New name"
                  }
                }
              },
              "example": {
                "description": "New description",
                "name": "New name"
              }
            }
          }
        },
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "proj-123"
          }
        }
      ]
    },
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "List projects",
        "description": "List projects",
        "operationId": "listProjects",
        "parameters": [
          {
            "name": "paginationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "50"
            }
          },
          {
            "name": "filter[name][begins]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "Project"
            }
          },
          {
            "name": "filter[createdAt][between][value1]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "date1"
            }
          },
          {
            "name": "filter[createdAt][between][value2]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "date2"
            }
          },
          {
            "name": "filter[dependency][eq]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "cc-1"
            }
          },
          {
            "name": "filter[status][lte]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "AVAILABLE"
            }
          },
          {
            "name": "sort[createdAt]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "asc"
            }
          },
          {
            "name": "sort[dependency]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "desc"
            }
          },
          {
            "name": "sort[name]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "desc"
            }
          },
          {
            "name": "sort[status]",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "asc"
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create Project",
        "description": "Create Project",
        "operationId": "createProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "costCenterId": {
                    "type": "string",
                    "example": "cc-exampleCostCenterId"
                  },
                  "description": {
                    "type": "string",
                    "example": "Example Project 1"
                  },
                  "name": {
                    "type": "string",
                    "example": "Example Project 1"
                  }
                }
              },
              "example": {
                "costCenterId": "cc-exampleCostCenterId",
                "description": "Example Project 1",
                "name": "Example Project 1"
              }
            }
          }
        },
        "responses": {
          "201": {}
        }
      }
    },
...
    "/projects/{projectId}/users": {
      "get": {
        "tags": ["projects"],
        "summary": "List Project Users By Role",
        "description": "List Project Users By Role",
        "operationId": "listProjectUsersByRole",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "ProjectAdmin"
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
    "/projects/{projectId}/users/{userId}/relationships": {
      "post": {
        "tags": ["projects"],
        "summary": "Add User To Project",
        "description": "Add User To Project",
        "operationId": "addUserToProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "example": "Researcher"
                  }
                }
              },
              "example": {
                "role": "Researcher"
              }
            }
          }
        },
        "responses": {
          "204": {}
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Remove User From Project",
        "description": "Remove User From Project",
        "operationId": "removeUserFromProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        },
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
...
    "/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Get User",
        "description": "Get User",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update User",
        "description": "Update User",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "status": {
                    "type": "string",
                    "example": "ACTIVE"
                  }
                }
              },
              "example": {
                "email": "john.doe@example.com",
                "firstName": "John",
                "lastName": "Doe",
                "status": "ACTIVE"
              }
            }
          }
        },
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
...
  },
...
}